+++ ouroboros-consensus-byron-test/src/Test/Consensus/Byron/Examples.hs
@@ -37,7 +37,7 @@ import           Ouroboros.Consensus.Config
 import           Ouroboros.Consensus.HeaderValidation
 import           Ouroboros.Consensus.Ledger.Abstract
 import           Ouroboros.Consensus.Ledger.Extended
-import qualified Ouroboros.Consensus.Mempool.TxLimits as TxLimits
+import qualified Ouroboros.Consensus.Mempool as Mempool
 import           Ouroboros.Consensus.NodeId
 import           Ouroboros.Consensus.Protocol.Abstract
 import           Ouroboros.Consensus.Protocol.PBFT
@@ -125,7 +125,7 @@ exampleBlock :: ByronBlock
 exampleBlock =
     forgeRegularBlock
       cfg
-      (TxLimits.mkOverrides TxLimits.noOverridesMeasure)
+      (Mempool.mkOverrides Mempool.noOverridesMeasure)
       (BlockNo 1)
       (SlotNo 1)
       (applyChainTick ledgerConfig (SlotNo 1) ledgerStateAfterEBB)
