+++ ouroboros-consensus-cardano/changelog.d/20230202_133339_javier.sagredo_reorganize_mempool.md
@@ -1,0 +1,22 @@
+<!--
+A new scriv changelog fragment.
+
+Uncomment the section that is right (remove the HTML comment wrapper).
+-->
+
+<!--
+### Patch
+
+- A bullet item for the Patch category.
+
+-->
+### Non-Breaking
+
+- Adapted to the new reorganization of Mempool modules in `ouroboros-consensus`.
+
+<!--
+### Breaking
+
+- A bullet item for the Breaking category.
+
+-->
