+++ ouroboros-consensus-shelley-test/test/Test/Consensus/Shelley/Coherence.hs
@@ -6,7 +6,7 @@ import           Data.Word (Word32)
 import           Test.Tasty
 import           Test.Tasty.QuickCheck
 
-import qualified Ouroboros.Consensus.Mempool.TxLimits as TxLimits
+import qualified Ouroboros.Consensus.Mempool.Capacity as MempoolCapacity
 
 import           Cardano.Ledger.Alonzo.Scripts (ExUnits, pointWiseExUnits)
 import           Test.Cardano.Ledger.Alonzo.Serialisation.Generators ()
@@ -16,15 +16,15 @@ import           Ouroboros.Consensus.Shelley.Ledger.Mempool (AlonzoMeasure (..),
 
 tests :: TestTree
 tests = testGroup "Shelley coherences" [
-      testProperty "TxLimits.<= uses pointWiseExUnits (<=)" leqCoherence
+      testProperty "MempoolCapacity.<= uses pointWiseExUnits (<=)" leqCoherence
     ]
 
--- | 'TxLimits.<=' and @'pointWiseExUnits' (<=)@ must agree
+-- | 'MempoolCapacity.<=' and @'pointWiseExUnits' (<=)@ must agree
 leqCoherence :: Word32 -> ExUnits -> ExUnits -> Property
 leqCoherence w eu1 eu2 =
     actual === expected
   where
-    inj eu = AlonzoMeasure (TxLimits.ByteSize w) (fromExUnits eu)
+    inj eu = AlonzoMeasure (MempoolCapacity.ByteSize w) (fromExUnits eu)
 
-    actual   = inj eu1 TxLimits.<= inj eu2
+    actual   = inj eu1 MempoolCapacity.<= inj eu2
     expected = pointWiseExUnits (<=) eu1 eu2
