+++ ouroboros-consensus/changelog.d/20230202_133330_javier.sagredo_reorganize_mempool.md
@@ -1,0 +1,22 @@
+<!--
+A new scriv changelog fragment.
+
+Uncomment the section that is right (remove the HTML comment wrapper).
+-->
+
+<!--
+### Patch
+
+- A bullet item for the Patch category.
+
+-->
+### Non-Breaking
+
+- Reorganization on the `Mempool` modules. Stub deprecated modules are in place
+  which should ensure that no code breaks downstream just yet. Clients should
+  directly import `Ouroboros.Consensus.Mempool`.
+
+### Breaking
+
+- Removed the `idx` type variable on the `Mempool` and `MempoolSnapshot`
+  datatypes in favour of using `TicketNo` always.
