+++ ouroboros-consensus/src/Ouroboros/Consensus/Mempool.hs
@@ -1,6 +1,59 @@
 -- | The mempool API and implementation.
-module Ouroboros.Consensus.Mempool (module X) where
+module Ouroboros.Consensus.Mempool (
+    -- * Mempool API
+    -- ** Mempool
+    Mempool (..)
+    -- ** Transaction adding
+  , MempoolAddTxResult (..)
+  , addLocalTxs
+  , addTxs
+  , isMempoolTxAdded
+  , isMempoolTxRejected
+  , mempoolTxAddedToMaybe
+    -- ** Ledger state to forge on top of
+  , ForgeLedgerState (..)
+    -- ** Mempool Snapshot
+  , MempoolSnapshot (..)
+    -- ** Re-exports
+  , TicketNo
+  , TxSizeInBytes
+  , zeroTicketNo
+    -- * Mempool capacity
+  , MempoolCapacityBytes (..)
+  , MempoolCapacityBytesOverride (..)
+  , computeMempoolCapacity
+    -- ** Mempool Size
+  , MempoolSize (..)
+    -- ** Transaction size
+  , ByteSize (..)
+  , TxLimits (..)
+    -- ** Restricting more strongly than the ledger's limits
+  , TxOverrides
+  , applyOverrides
+  , getOverrides
+  , mkOverrides
+  , noOverridesMeasure
+    -- * Mempool initialization
+  , openMempool
+  , openMempoolWithoutSyncThread
+    -- * ChainDB interface
+  , LedgerInterface (..)
+  , chainDBLedgerInterface
+    -- * Trace
+  , TraceEventMempool (..)
+  ) where
 
-import           Ouroboros.Consensus.Mempool.API as X
-import           Ouroboros.Consensus.Mempool.Impl as X
-import           Ouroboros.Consensus.Mempool.TxSeq as X (TicketNo)
+import           Ouroboros.Consensus.Mempool.API (ForgeLedgerState (..),
+                     Mempool (..), MempoolAddTxResult (..),
+                     MempoolSnapshot (..), TicketNo, TxSizeInBytes, addLocalTxs,
+                     addTxs, isMempoolTxAdded, isMempoolTxRejected,
+                     mempoolTxAddedToMaybe, zeroTicketNo)
+import           Ouroboros.Consensus.Mempool.Capacity (ByteSize (..),
+                     MempoolCapacityBytes (..),
+                     MempoolCapacityBytesOverride (..), MempoolSize (..),
+                     TxLimits (..), TxOverrides (..), applyOverrides,
+                     computeMempoolCapacity, mkOverrides, noOverridesMeasure)
+import           Ouroboros.Consensus.Mempool.Impl.Common (LedgerInterface (..),
+                     TraceEventMempool (..), chainDBLedgerInterface)
+import           Ouroboros.Consensus.Mempool.Init (openMempool,
+                     openMempoolWithoutSyncThread)
